cpufreq_darkness: Made load_change_callback() inline
As mentioned in one of the earlier commits, GCC triggers a fatal warning when an unused function is present in the file. A fix for the same is to make the function inline, or delete it if it isn't being used. I have opted for the former in this case.